pyproject.toml
[project]
name = "mcp-blackboard"
version = "0.1.0"
description = "A mcp server for providing a blackboard interface to the mcp agent"
authors = [
    { name = "Kwesi Apponsah", email = "kwesparry@gmail.com" },
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "apscheduler>=3.11.0",
    "dotenv>=0.9.9",
    "fsspec[abfs,fuse,gcs,s3,sftp,smb]>=2025.3.2",
    "jinja2>=3.1.6",
    "markitdown[all]>=0.1.1",
    "mcp[cli]>=1.6.0",
    "openai>=1.76.0",
    "pydantic>=2.11.3",
    "pydantic-settings>=2.9.1",
    "pyyaml>=6.0.2",
    "redis>=5.2.1",
    "rich>=14.0.0",
    "tenacity>=9.1.2",
    "types-pyyaml>=6.0.12.20250516",
]

[dependency-groups]
dev = [
    "coverage>=7.8.0",
    "mypy>=1.15.0",
    "pre-commit>=4.2.0",
    "pytest>=8.3.5",
    "pytest-asyncio>=0.26.0",
    "pytest-cov>=6.1.1",
    "ruff>=0.11.7",
]
infra = [
    "pulumi>=3.167.0",
    "pulumi-azure-native>=3.3.0",
]


[tool.ruff]
lint.select = [
    "E",  # pycodestyle errors
    "W",  # pycodestyle warnings
    "F",  # pyflakes
    "I",  # isort
    "B",  # flake8-bugbear
    "C4", # flake8-comprehensions
    "UP", # pyupgrade,
]

[tool.ruff.lint.pydocstyle]
convention = "google"

[tool.ruff.lint.per-file-ignores]
"examples/**/*.py" = ["E501"]

[tool.mypy]
strict = true
follow_imports = "silent"
warn_redundant_casts = true
warn_unused_ignores = false
disallow_any_generics = true
no_implicit_reexport = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
disallow_untyped_calls = true
explicit_package_bases = true
disallow_untyped_decorators = false
mypy_path = "src"
plugins = [
  "pydantic.mypy"
]

[tool.pydantic-mypy]
init_forbid_extra = true
init_typed = true
warn_required_dynamic_aliases = true

[tool.coverage.run]
source = ["src"]
omit = [
    "*/__init__.py",
    "src/*/__init__.py",
    "src/**/__init__.py",
]


[tool.coverage.report]
show_missing = true
sort = "-Cover"
exclude_also = [
    # This is only executed while typechecking
    "if TYPE_CHECKING:",
    "@abc.abstractmethod",
    "raise NotImplementedError",
    "logger.debug",
]

[tool.pytest.ini_options]
addopts = [
    "--pdb",
    "--tb=short",
    "--disable-warnings",
    "--maxfail=1",
]
testpaths = [
    "tests"
]
pythonpath = [
    "src"
]

asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "session"
filterwarnings = [
    # This is a warning that is expected to happen: we have an async filter that raises an exception
    "ignore:DeprecatedWarning:RuntimeWarning:UserWarning",
]
